refactor: updater package has been mode into the background_services package.
